feat(review): Implement average rating calculation for tours
Added a static method calcAverageRatings in the Review model to compute the average rating and rating quantity for a specific tour.
Updated the tour document's ratingQuantity and ratingAverage fields based on the aggregation of reviews.
Implemented post('save') middleware to trigger the recalculation of ratings each time a review is added.
Added pre and post middleware for findOneAndUpdate and findOneAndDelete to ensure that tour ratings are updated when reviews are modified or deleted.
Improved query execution by ensuring that the review document is accessed only once during the middleware lifecycle.